Conflicts:
	client/client_daemon.py
	client/test_client_daemon.py

+++ client/connection_manager.py
@@ -19,12 +19,17 @@
 # - DELETE /shares/<root_path>/<user> - elimina l’utente dallo share
 
 import requests
+import urllib
 import json
 import os
 import logging
+import keyring
 
 
 class ConnectionManager(object):
+    # This is the char filter for url encoder, this list of char aren't translated in percent style
+    ENCODER_FILTER = '+/: '
+
     EXCEPTIONS_CATCHED = (requests.HTTPError,
                           requests.exceptions.ConnectionError,
                           requests.exceptions.MissingSchema,
@@ -62,7 +67,7 @@ class ConnectionManager(object):
         :param cfg: Dictionary where is contained the configuration
         """
         self.cfg = cfg
-        self.auth = (self.cfg.get('user', None), self.cfg.get('pass', None))
+        self.auth = (self.cfg.get('user'), keyring.get_password('PyBox', self.cfg.get('user', '')))
 
         # example of self.base_url = 'http://localhost:5000/API/V1/'
         self.base_url = ''.join([self.cfg['server_address'], self.cfg['api_suffix']])
@@ -78,27 +83,41 @@ class ConnectionManager(object):
         except AttributeError:
             self._default(method_name)
 
+    def do_login(self, data):
+        url = self.files_url
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
+        user = data[0]
+        password = data[1]
+        self.logger.info('{}: URL: {} - DATA: {} '.format('do_login', url, data))
+        try:
+            r = requests.get(encoded_url, auth=(user, password))
+            r.raise_for_status()
+            return {'content': 'User authenticated', 'successful': True}
+        except ConnectionManager.EXCEPTIONS_CATCHED as e:
+            self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_login', url, e))
+            return {'content': 'Impossible to login, user unauthorized.', 'successful': False}
+
     def do_register(self, data):
         """
         Send registration user request
         """
         req = {'password': data[1]}
         url = ''.join([self.users_url, data[0]])
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
         self.logger.info('do_register: URL: {} - DATA: {} '.format(url, data))
 
         try:
-            r = requests.post(url, data=req)
+            r = requests.post(encoded_url, data=req)
             # i must check before raise_for_status to not destroy response
             if r.status_code == 403:
                 return {'improvements': json.loads(r.text), 'successful': False}
             elif r.status_code == 409:
                 return {'content': 'Error! User already existent!', 'successful': False}
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('do_register: URL: {} - EXCEPTION_CATCHED: {} '.format(url, e))
             return {'content': 'Error during registration:\n{}'.format(e), 'successful': False}
-        else:
-            return {'content': json.loads(r.text), 'successful': True}
 
     def do_activate(self, data):
         """
@@ -106,20 +125,20 @@ class ConnectionManager(object):
         """
         req = {'activation_code': data[1]}
         url = ''.join([self.users_url, data[0]])
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
         self.logger.info('do_activate: URL: {} - DATA: {} '.format(url, data))
 
         try:
-            r = requests.put(url, data=req)
+            r = requests.put(encoded_url, data=req)
             if r.status_code == 404:
                 return {'content': 'Error! Impossible to activate user! Unexistent user!', 'successful': False}
             elif r.status_code == 409:
                 return {'content': 'Error! Impossible to activate user! User already activated!', 'successful': False}
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('do_activate: URL: {} - EXCEPTION_CATCHED: {} '.format(url, e))
-        else:
-            return {'content': json.loads(r.text), 'successful': True}
-        return {'content': 'Error during activation:\n{}'.format(e), 'successful': False}
+            return {'content': 'Error during activation:\n{}'.format(e), 'successful': False}
 
     def do_reqrecoverpass(self, data):
         """
@@ -127,13 +146,13 @@ class ConnectionManager(object):
         """
         mail = data
         url = '{}{}/reset'.format(self.users_url, mail)
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
         try:
-            r = requests.post(url)
+            r = requests.post(encoded_url)
             r.raise_for_status()
+            return r.text
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('do_reqrecoverpass: URL: {} - EXCEPTION_CATCHED: {}'.format(url, e))
-        else:
-            return r.text
 
     def do_recoverpass(self, data):
         """
@@ -141,15 +160,15 @@ class ConnectionManager(object):
         """
         mail, recoverpass_code, new_password = data
         url = '{}{}'.format(self.users_url, mail)
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
         try:
-            r = requests.put(url,
+            r = requests.put(encoded_url,
                              data={'password': new_password,
                                    'recoverpass_code': recoverpass_code})
             r.raise_for_status()
+            return r.text
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('do_recoverpass: URL: {} - EXCEPTION_CATCHED: {}'.format(url, e))
-        else:
-            return r.text
 
     # shares
 
@@ -208,57 +227,59 @@ class ConnectionManager(object):
 
     def do_download(self, data):
         url = ''.join([self.files_url, data['filepath']])
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
         self.logger.info('{}: URL: {} - DATA: {} '.format('do_download', url, data))
         try:
-            r = requests.get(url, auth=self.auth)
+            r = requests.get(encoded_url, auth=self.auth)
             r.raise_for_status()
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_download', url, e))
-        else:
-            filepath = os.path.join(self.cfg['sharing_path'], data['filepath'])
-            dirpath, filename = os.path.split(filepath)
-            if not os.path.isdir(dirpath):
-                # Create all missing directories
-                os.makedirs(dirpath)
+            return {'content': 'Failed to download file from server.\n'
+                               'Path: {}\nError: {}'.format(data['filepath'], e),
+                    'successful': False}
+        filepath = os.path.join(self.cfg['sharing_path'], data['filepath'])
+        dirpath, filename = os.path.split(filepath)
+        # Create all missing directories
+        if not os.path.isdir(dirpath):
+            os.makedirs(dirpath)
+        if not os.path.exists(filepath):
             with open(filepath, 'wb') as f:
                 f.write(r.content)
-            return True
-        return False
+                return {'successful': True}
+        else:
+            return {'content': 'Warning! Download of file already existent! Operation Aborted.', 'successful': False}
 
     def do_upload(self, data):
         filepath = os.path.join(self.cfg['sharing_path'], data['filepath'])
         url = ''.join([self.files_url, data['filepath']])
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
         self.logger.info('{}: URL: {} - DATA: {} '.format('do_upload', url, data))
         _file = {'file': (open(filepath, 'rb'))}
-
         try:
-            r = requests.post(url, auth=self.auth, files=_file, data={'md5': data['md5']})
+            r = requests.post(encoded_url, auth=self.auth, files=_file, data={'md5': data['md5']})
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_upload', url, e))
-        else:
-            event_timestamp = json.loads(r.text)
-
-            return event_timestamp
-        return False
+            return {'content': 'Failed to upload file to the server.\n'
+                               'Path: {}\nError: {}'.format(data['filepath'], e),
+                    'successful': False}
 
     def do_modify(self, data):
         filepath = os.path.join(self.cfg['sharing_path'], data['filepath'])
         url = ''.join([self.files_url, data['filepath']])
-
+        encoded_url = urllib.quote(url, ConnectionManager.ENCODER_FILTER)
         self.logger.info('{}: URL: {} - DATA: {} '.format('do_modify', url, data))
-
         _file = {'file': (open(filepath, 'rb'))}
         try:
-            r = requests.put(url, auth=self.auth, files=_file, data={'md5': data['md5']})
+            r = requests.put(encoded_url, auth=self.auth, files=_file, data={'md5': data['md5']})
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_modify', url, e))
-        else:
-            event_timestamp = json.loads(r.text)
-
-            return event_timestamp
-        return False
+            return {'content': 'Failed to modify file on server.\n'
+                               'Path: {}\nError: {}'.format(data['filepath'], e),
+                    'successful': False}
 
     # actions:
 
@@ -269,13 +290,12 @@ class ConnectionManager(object):
         try:
             r = requests.post(url, auth=self.auth, data=d)
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_move', url, e))
-        else:
-            event_timestamp = json.loads(r.text)
-
-            return event_timestamp
-        return False
+            return {'content': 'Failed to move file on server.\n'
+                               'Src path: {}\nDest Path: {}\nError: {}'.format(data['src'], data['dst'], e),
+                    'successful': False}
 
     def do_delete(self, data):
         url = ''.join([self.actions_url, 'delete'])
@@ -284,12 +304,12 @@ class ConnectionManager(object):
         try:
             r = requests.post(url, auth=self.auth, data=d)
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_delete', url, e))
-        else:
-            event_timestamp = json.loads(r.text)
-            return event_timestamp
-        return False
+            return {'content': 'Failed to delete file on server.\n'
+                               'Path: {}\nError: {}'.format(data['filepath'], e),
+                    'successful': False}
 
     def do_copy(self, data):
         url = ''.join([self.actions_url, 'copy'])
@@ -298,28 +318,24 @@ class ConnectionManager(object):
         try:
             r = requests.post(url, auth=self.auth, data=d)
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
             self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_copy', url, e))
-        else:
-            event_timestamp = json.loads(r.text)
-
-            return event_timestamp
-        return False
+            return {'content': 'Failed to copy file on server.\n'
+                               'Src path: {}\nDest Path: {}\nError: {}'.format(data['src'], data['dst'], e),
+                    'successful': False}
 
     def do_get_server_snapshot(self, data):
         url = self.files_url
-
         self.logger.info('{}: URL: {} - DATA: {} '.format('do_get_server_snapshot', url, data))
 
         try:
             r = requests.get(url, auth=self.auth)
             r.raise_for_status()
+            return {'content': r.json(), 'successful': True}
         except ConnectionManager.EXCEPTIONS_CATCHED as e:
-
             self.logger.error('{}: URL: {} - EXCEPTION_CATCHED: {} '.format('do_get_server_snapshot', url, e))
-
-        else:
-            return json.loads(r.text)
+            return {'content': 'Failed to get server snapshot, maybe server down?\nError: {}'.format(e), 'successful': False}
 
     def _default(self, method):
         print 'Received Unknown Command:', method
